Paper-trading effectiveness pack: 5 fixes targeting the $2k/4w P&L target
After re-framing from production-readiness to paper-trading effectiveness,
five fixes shipped that move the P&L needle directly:

1. **Fees + slippage calibration** (config.json)
   fee 0.0025 → 0.005 to match Coinbase Advanced blended taker/maker.
   Strategy backtests no longer flatter themselves at half the real cost.

2. **B-22 regime-stability gate** (FreqAIMeanRevV1.py + config.json)
   New regime_min_stable_hours: 2.0 gate blocks entries within N hours
   of an HMM regime flip. Today's 3-for-3 paper losses all entered
   minutes after a flip and stopped out when the flip reversed. Column
   missing or regime-source down → duration=0.0 → gate blocks (safer
   default). BollingerRSI MR path is intentionally exempt — it WANTS
   post-flip dips.

3. **Stocks TFT inference gate** (market_open.py + tft_stock.py)
   New _tft_gate() runs the trained TFT model on each BUY candidate
   after LLM confidence/R:R checks pass. Vetoes when DOWN > UP with
   ≥0.05 confidence, or when UP < 0.40. Fails open on infrastructure
   errors. Tested against 8 tickers — model correctly vetoes 6/8 in
   today's chop. Also fixed dtype bug in predict_direction (np.zeros
   defaulted to float64, model expects float32) that silently broke
   inference for any ticker without trained-time norms.

4. **Stocks notifier wiring** (notify.py shim + 3 phase files)
   New shark/notify.py shim re-exposes user_data/modules/notifier.py
   singleton from inside the stocks codebase. Wired into
   market_open._execute (BUY signal → Slack/Telegram), daily_summary
   (EOD summary), execution/orders (order placement failure). Operator
   no longer has to tail logs to learn the stocks side did something.

5. **Circuit-breaker false-alarm fix** (unified_risk.py + ops_spa.js)
   stocks_data_stale + stocks_data_untrusted were displayed as TRIPPED
   on the dashboard every evening + weekend because wheel_snapshot
   cron only runs Mon-Fri 9-16 ET. Backend now exposes market_open_now
   in the RiskStatus payload; frontend hides the rows ("market closed
   — gate inactive") when market is closed, mirroring the backend
   breaker's existing market_open_now gate.

Plus cache-bust cutover18 → cutover19 to ship the FE change.

+++ stocks/shark/ml/tft_stock.py
@@ -330,8 +330,12 @@ def predict_direction(
     model, norms, payload, _cols = _load_for_inference(weights_path)
     mu, sd = norms.get(ticker, (None, None))
     if mu is None:
-        mu = np.zeros(feature_window.shape[1])
-        sd = np.ones(feature_window.shape[1])
+        # New ticker absent from training-time norms — neutral z-score.
+        # MUST be float32 or the model.forward() linear layer rejects with
+        # "expected mat1 and mat2 to have the same dtype, but got: double != float"
+        # because np.zeros defaults to float64.
+        mu = np.zeros(feature_window.shape[1], dtype=np.float32)
+        sd = np.ones(feature_window.shape[1], dtype=np.float32)
     else:
         mu = np.asarray(mu, dtype=np.float32)
         sd = np.asarray(sd, dtype=np.float32)
